refactor: Repository インターフェースを各ファイルに分散
